requirements.txt
fastapi
uvicorn[standard]
jinja2
aiofiles
gunicorn
python-dotenv>=1.0.1
substrate-interface>=1.7.11
tenacity>=8.2.3
